requirements.txt
amqp==2.6.1
asgiref==3.2.10
astroid==2.4.2
billiard==3.6.3.0
celery==4.4.7
celery-progress==0.0.12
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distance==0.1.3
Django==3.1.1
django-celery-results==1.2.1
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
idna==2.10
ImageHash==4.2.0
isort==5.5.2
joblib==0.16.0
jsonfield==3.1.0
jsonmerge
kombu==4.6.11
lazy-object-proxy==1.4.3
mccabe==0.6.1
mod-wsgi==4.7.1
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
pandas==1.1.2
Pillow==8.1.2
postgres
psaw==0.1.0
psycopg2-binary==2.8.6
pycountry
pylint==2.6.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
regex
requests==2.24.0
requests-oauthlib==1.3.0
scipy==1.6.2
six==1.15.0
SQLAlchemy==1.3.19
sqlparse==0.3.1
toml==0.10.1
tqdm
tweepy==3.9.0
typed-ast==1.4.2
urllib3==1.25.10
vine==1.3.0
wget==3.2
wrapt==1.12.1
zipp==3.4.1
sklearn
